Remix routes in storybook poc

+++ tests/db-utils.ts
@@ -1,8 +1,8 @@
-import fs from 'node:fs'
 import { faker } from '@faker-js/faker'
 import { type PrismaClient } from '@prisma/client'
 import bcrypt from 'bcryptjs'
 import { UniqueEnforcer } from 'enforce-unique'
+import path from 'path'
 
 const uniqueUsernameEnforcer = new UniqueEnforcer()
 
@@ -108,10 +108,11 @@ export async function img({
 	altText?: string
 	filepath: string
 }) {
+	const url = path.relative('tests/fixtures', filepath)
 	return {
 		altText,
 		contentType: filepath.endsWith('.png') ? 'image/png' : 'image/jpeg',
-		blob: await fs.promises.readFile(filepath),
+		blob: await (await fetch(url)).arrayBuffer(),
 	}
 }
 
@@ -119,14 +120,19 @@ export async function cleanupDb(prisma: PrismaClient) {
 	const tables = await prisma.$queryRaw<
 		{ name: string }[]
 	>`SELECT name FROM sqlite_master WHERE type='table' AND name NOT LIKE 'sqlite_%' AND name NOT LIKE '_prisma_migrations';`
-
-	await prisma.$transaction([
-		// Disable FK constraints to avoid relation conflicts during deletion
-		prisma.$executeRawUnsafe(`PRAGMA foreign_keys = OFF`),
-		// Delete all rows from each table, preserving table structures
-		...tables.map(({ name }) =>
-			prisma.$executeRawUnsafe(`DELETE from "${name}"`),
-		),
-		prisma.$executeRawUnsafe(`PRAGMA foreign_keys = ON`),
-	])
+	if ('$getInternalState' in prisma) {
+		const data = prisma.$getInternalState()
+		for (var member in data) data[member] = []
+	} else {
+		await prisma.$transaction([
+			// Disable FK constraints to avoid relation conflicts during deletion
+			prisma.$executeRawUnsafe(`PRAGMA foreign_keys = OFF`),
+			// Delete all rows from each table, preserving table structures
+			...(tables ?? []).map(({ name }) =>
+				prisma.$executeRawUnsafe(`DELETE from "${name}"`),
+			),
+			prisma.$executeRawUnsafe(`PRAGMA foreign_keys = ON`),
+		])
+	}
 }
+
